ZCS-3702:Some Spaces removed in RFC 2047 encoded subject
[Bug]
When the subject contains the "encoded word" ("=?<charset>?<decode>?...?="
sequence), the space before and after it were removed.

[Root cause]
Previously, the decoder was implemented to comply with the
description of the RFC 2047 Section 6.2, and removed any white space
before and after the "encoded-word", but it should have removed
the 'linear-white-space' only between the 'encoded-word', not every ones.
```
6.2
   When displaying a particular header field that contains multiple
   'encoded-word's, any 'linear-white-space' that separates a pair of
   adjacent 'encoded-word's is ignored.  (This is to allow the use of
   multiple 'encoded-word's to represent long strings of unencoded text,
   without having to separate 'encoded-word's where spaces occur in the
   unencoded text.)
```
[Fix]
* Rewrote the decoder "decode()"
* Fixed the unit test cases: removed the intentionally added white spaces.
  For example, the space before the "is" is included in the encoded-word.
  ```
  "こんにちは" is "Hello" in Japanese
  ```